requirements.txt
pandas
plotly_express
streamlit
plotly.graph_objects